2009-01-20 Gonzalo Paniagua Javier <gonzalo@novell.com>
	* IPAddress.cs: handle sign extension for the high byte when using the
	byte[] ctor for an IPv4 address.
	Fixes bug #467472.


svn path=/trunk/mcs/; revision=123856

+++ mcs/class/System/System.Net/IPAddress.cs
@@ -153,7 +153,7 @@ namespace System.Net {
 				m_Family = AddressFamily.InterNetworkV6;
 				m_ScopeId = 0;
 			} else {
-				m_Address = (address [3] << 24) + (address [2] << 16) +
+				m_Address = ((uint) address [3] << 24) + (address [2] << 16) +
 					(address [1] << 8) + address [0];
 			}
 		}
